+++ apis/appmesh/v1beta1/zz_gatewayroute_types.go
@@ -64,7 +64,7 @@ type ActionRewriteParameters struct {
 type ActionTargetInitParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []TargetVirtualServiceInitParameters `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -73,7 +73,7 @@ type ActionTargetInitParameters struct {
 type ActionTargetObservation struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []TargetVirtualServiceObservation `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -83,7 +83,7 @@ type ActionTargetParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	// +kubebuilder:validation:Optional
@@ -276,7 +276,7 @@ type HTTPRouteActionParameters struct {
 type HTTPRouteActionTargetInitParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []ActionTargetVirtualServiceInitParameters `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -285,7 +285,7 @@ type HTTPRouteActionTargetInitParameters struct {
 type HTTPRouteActionTargetObservation struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []ActionTargetVirtualServiceObservation `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -295,7 +295,7 @@ type HTTPRouteActionTargetParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	// +kubebuilder:validation:Optional
@@ -352,7 +352,7 @@ type HTTPRouteMatchInitParameters struct {
 	Path []MatchPathInitParameters `json:"path,omitempty" tf:"path,omitempty"`
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	Prefix *string `json:"prefix,omitempty" tf:"prefix,omitempty"`
@@ -373,7 +373,7 @@ type HTTPRouteMatchObservation struct {
 	Path []MatchPathObservation `json:"path,omitempty" tf:"path,omitempty"`
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	Prefix *string `json:"prefix,omitempty" tf:"prefix,omitempty"`
@@ -398,7 +398,7 @@ type HTTPRouteMatchParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	// +kubebuilder:validation:Optional
@@ -596,7 +596,7 @@ type Http2RouteMatchInitParameters struct {
 	Path []PathInitParameters `json:"path,omitempty" tf:"path,omitempty"`
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	Prefix *string `json:"prefix,omitempty" tf:"prefix,omitempty"`
@@ -617,7 +617,7 @@ type Http2RouteMatchObservation struct {
 	Path []PathObservation `json:"path,omitempty" tf:"path,omitempty"`
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	Prefix *string `json:"prefix,omitempty" tf:"prefix,omitempty"`
@@ -642,7 +642,7 @@ type Http2RouteMatchParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Specified beginning characters to rewrite.
 	// +kubebuilder:validation:Optional
@@ -803,7 +803,7 @@ type MatchHostnameParameters struct {
 type MatchInitParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Fully qualified domain name for the service to match from the request.
 	ServiceName *string `json:"serviceName,omitempty" tf:"service_name,omitempty"`
@@ -812,7 +812,7 @@ type MatchInitParameters struct {
 type MatchObservation struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Fully qualified domain name for the service to match from the request.
 	ServiceName *string `json:"serviceName,omitempty" tf:"service_name,omitempty"`
@@ -822,7 +822,7 @@ type MatchParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Fully qualified domain name for the service to match from the request.
 	// +kubebuilder:validation:Optional
@@ -909,30 +909,30 @@ type MatchQueryParameterParameters struct {
 type MatchRangeInitParameters struct {
 
 	// End of the range.
-	End *float64 `json:"end,omitempty" tf:"end,omitempty"`
+	End *int64 `json:"end,omitempty" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
-	Start *float64 `json:"start,omitempty" tf:"start,omitempty"`
+	Start *int64 `json:"start,omitempty" tf:"start,omitempty"`
 }
 
 type MatchRangeObservation struct {
 
 	// End of the range.
-	End *float64 `json:"end,omitempty" tf:"end,omitempty"`
+	End *int64 `json:"end,omitempty" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
-	Start *float64 `json:"start,omitempty" tf:"start,omitempty"`
+	Start *int64 `json:"start,omitempty" tf:"start,omitempty"`
 }
 
 type MatchRangeParameters struct {
 
 	// End of the range.
 	// +kubebuilder:validation:Optional
-	End *float64 `json:"end" tf:"end,omitempty"`
+	End *int64 `json:"end" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
 	// +kubebuilder:validation:Optional
-	Start *float64 `json:"start" tf:"start,omitempty"`
+	Start *int64 `json:"start" tf:"start,omitempty"`
 }
 
 type PathInitParameters struct {
@@ -1044,30 +1044,30 @@ type QueryParameterParameters struct {
 type RangeInitParameters struct {
 
 	// End of the range.
-	End *float64 `json:"end,omitempty" tf:"end,omitempty"`
+	End *int64 `json:"end,omitempty" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
-	Start *float64 `json:"start,omitempty" tf:"start,omitempty"`
+	Start *int64 `json:"start,omitempty" tf:"start,omitempty"`
 }
 
 type RangeObservation struct {
 
 	// End of the range.
-	End *float64 `json:"end,omitempty" tf:"end,omitempty"`
+	End *int64 `json:"end,omitempty" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
-	Start *float64 `json:"start,omitempty" tf:"start,omitempty"`
+	Start *int64 `json:"start,omitempty" tf:"start,omitempty"`
 }
 
 type RangeParameters struct {
 
 	// End of the range.
 	// +kubebuilder:validation:Optional
-	End *float64 `json:"end" tf:"end,omitempty"`
+	End *int64 `json:"end" tf:"end,omitempty"`
 
 	// (Requited) Start of the range.
 	// +kubebuilder:validation:Optional
-	Start *float64 `json:"start" tf:"start,omitempty"`
+	Start *int64 `json:"start" tf:"start,omitempty"`
 }
 
 type RewriteHostnameInitParameters struct {
@@ -1159,7 +1159,7 @@ type SpecInitParameters struct {
 	Http2Route []Http2RouteInitParameters `json:"http2Route,omitempty" tf:"http2_route,omitempty"`
 
 	// Priority for the gateway route, between 0 and 1000.
-	Priority *float64 `json:"priority,omitempty" tf:"priority,omitempty"`
+	Priority *int64 `json:"priority,omitempty" tf:"priority,omitempty"`
 }
 
 type SpecObservation struct {
@@ -1174,7 +1174,7 @@ type SpecObservation struct {
 	Http2Route []Http2RouteObservation `json:"http2Route,omitempty" tf:"http2_route,omitempty"`
 
 	// Priority for the gateway route, between 0 and 1000.
-	Priority *float64 `json:"priority,omitempty" tf:"priority,omitempty"`
+	Priority *int64 `json:"priority,omitempty" tf:"priority,omitempty"`
 }
 
 type SpecParameters struct {
@@ -1193,13 +1193,13 @@ type SpecParameters struct {
 
 	// Priority for the gateway route, between 0 and 1000.
 	// +kubebuilder:validation:Optional
-	Priority *float64 `json:"priority,omitempty" tf:"priority,omitempty"`
+	Priority *int64 `json:"priority,omitempty" tf:"priority,omitempty"`
 }
 
 type TargetInitParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []VirtualServiceInitParameters `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -1208,7 +1208,7 @@ type TargetInitParameters struct {
 type TargetObservation struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	VirtualService []VirtualServiceObservation `json:"virtualService,omitempty" tf:"virtual_service,omitempty"`
@@ -1218,7 +1218,7 @@ type TargetParameters struct {
 
 	// The port number that corresponds to the target for Virtual Service provider port. This is required when the provider (router or node) of the Virtual Service has multiple listeners.
 	// +kubebuilder:validation:Optional
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Virtual service gateway route target.
 	// +kubebuilder:validation:Optional
