+++ apis/cloudfront/v1beta1/zz_cachepolicy_types.go
@@ -19,13 +19,13 @@ type CachePolicyInitParameters struct {
 	Comment *string `json:"comment,omitempty" tf:"comment,omitempty"`
 
 	// The default amount of time, in seconds, that you want objects to stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
-	DefaultTTL *float64 `json:"defaultTtl,omitempty" tf:"default_ttl,omitempty"`
+	DefaultTTL *int64 `json:"defaultTtl,omitempty" tf:"default_ttl,omitempty"`
 
 	// The maximum amount of time, in seconds, that objects stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
-	MaxTTL *float64 `json:"maxTtl,omitempty" tf:"max_ttl,omitempty"`
+	MaxTTL *int64 `json:"maxTtl,omitempty" tf:"max_ttl,omitempty"`
 
 	// The minimum amount of time, in seconds, that you want objects to stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
-	MinTTL *float64 `json:"minTtl,omitempty" tf:"min_ttl,omitempty"`
+	MinTTL *int64 `json:"minTtl,omitempty" tf:"min_ttl,omitempty"`
 
 	// A unique name to identify the cache policy.
 	Name *string `json:"name,omitempty" tf:"name,omitempty"`
@@ -40,7 +40,7 @@ type CachePolicyObservation struct {
 	Comment *string `json:"comment,omitempty" tf:"comment,omitempty"`
 
 	// The default amount of time, in seconds, that you want objects to stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
-	DefaultTTL *float64 `json:"defaultTtl,omitempty" tf:"default_ttl,omitempty"`
+	DefaultTTL *int64 `json:"defaultTtl,omitempty" tf:"default_ttl,omitempty"`
 
 	// The current version of the cache policy.
 	Etag *string `json:"etag,omitempty" tf:"etag,omitempty"`
@@ -49,10 +49,10 @@ type CachePolicyObservation struct {
 	ID *string `json:"id,omitempty" tf:"id,omitempty"`
 
 	// The maximum amount of time, in seconds, that objects stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
-	MaxTTL *float64 `json:"maxTtl,omitempty" tf:"max_ttl,omitempty"`
+	MaxTTL *int64 `json:"maxTtl,omitempty" tf:"max_ttl,omitempty"`
 
 	// The minimum amount of time, in seconds, that you want objects to stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
-	MinTTL *float64 `json:"minTtl,omitempty" tf:"min_ttl,omitempty"`
+	MinTTL *int64 `json:"minTtl,omitempty" tf:"min_ttl,omitempty"`
 
 	// A unique name to identify the cache policy.
 	Name *string `json:"name,omitempty" tf:"name,omitempty"`
@@ -69,15 +69,15 @@ type CachePolicyParameters struct {
 
 	// The default amount of time, in seconds, that you want objects to stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
 	// +kubebuilder:validation:Optional
-	DefaultTTL *float64 `json:"defaultTtl,omitempty" tf:"default_ttl,omitempty"`
+	DefaultTTL *int64 `json:"defaultTtl,omitempty" tf:"default_ttl,omitempty"`
 
 	// The maximum amount of time, in seconds, that objects stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
 	// +kubebuilder:validation:Optional
-	MaxTTL *float64 `json:"maxTtl,omitempty" tf:"max_ttl,omitempty"`
+	MaxTTL *int64 `json:"maxTtl,omitempty" tf:"max_ttl,omitempty"`
 
 	// The minimum amount of time, in seconds, that you want objects to stay in the CloudFront cache before CloudFront sends another request to the origin to see if the object has been updated.
 	// +kubebuilder:validation:Optional
-	MinTTL *float64 `json:"minTtl,omitempty" tf:"min_ttl,omitempty"`
+	MinTTL *int64 `json:"minTtl,omitempty" tf:"min_ttl,omitempty"`
 
 	// A unique name to identify the cache policy.
 	// +kubebuilder:validation:Optional
