+++ apis/lambda/v1beta1/zz_layerversionpermission_types.go
@@ -31,7 +31,7 @@ type LayerVersionPermissionInitParameters struct {
 	StatementID *string `json:"statementId,omitempty" tf:"statement_id,omitempty"`
 
 	// Version of Lambda Layer, which you want to grant access to. Note: permissions only apply to a single version of a layer.
-	VersionNumber *float64 `json:"versionNumber,omitempty" tf:"version_number,omitempty"`
+	VersionNumber *int64 `json:"versionNumber,omitempty" tf:"version_number,omitempty"`
 }
 
 type LayerVersionPermissionObservation struct {
@@ -61,7 +61,7 @@ type LayerVersionPermissionObservation struct {
 	StatementID *string `json:"statementId,omitempty" tf:"statement_id,omitempty"`
 
 	// Version of Lambda Layer, which you want to grant access to. Note: permissions only apply to a single version of a layer.
-	VersionNumber *float64 `json:"versionNumber,omitempty" tf:"version_number,omitempty"`
+	VersionNumber *int64 `json:"versionNumber,omitempty" tf:"version_number,omitempty"`
 }
 
 type LayerVersionPermissionParameters struct {
@@ -93,7 +93,7 @@ type LayerVersionPermissionParameters struct {
 
 	// Version of Lambda Layer, which you want to grant access to. Note: permissions only apply to a single version of a layer.
 	// +kubebuilder:validation:Optional
-	VersionNumber *float64 `json:"versionNumber,omitempty" tf:"version_number,omitempty"`
+	VersionNumber *int64 `json:"versionNumber,omitempty" tf:"version_number,omitempty"`
 }
 
 // LayerVersionPermissionSpec defines the desired state of LayerVersionPermission
