+++ apis/secretsmanager/v1beta1/zz_secret_types.go
@@ -54,7 +54,7 @@ type RotationRulesInitParameters struct {
 type RotationRulesObservation struct {
 
 	// Specifies the number of days between automatic scheduled rotations of the secret.
-	AutomaticallyAfterDays *float64 `json:"automaticallyAfterDays,omitempty" tf:"automatically_after_days,omitempty"`
+	AutomaticallyAfterDays *int64 `json:"automaticallyAfterDays,omitempty" tf:"automatically_after_days,omitempty"`
 
 	Duration *string `json:"duration,omitempty" tf:"duration,omitempty"`
 
@@ -76,7 +76,7 @@ type SecretInitParameters struct {
 	Name *string `json:"name,omitempty" tf:"name,omitempty"`
 
 	// Number of days that AWS Secrets Manager waits before it can delete the secret. This value can be 0 to force deletion without recovery or range from 7 to 30 days. The default value is 30.
-	RecoveryWindowInDays *float64 `json:"recoveryWindowInDays,omitempty" tf:"recovery_window_in_days,omitempty"`
+	RecoveryWindowInDays *int64 `json:"recoveryWindowInDays,omitempty" tf:"recovery_window_in_days,omitempty"`
 
 	// Configuration block to support secret replication. See details below.
 	Replica []ReplicaInitParameters `json:"replica,omitempty" tf:"replica,omitempty"`
@@ -109,7 +109,7 @@ type SecretObservation struct {
 	Policy *string `json:"policy,omitempty" tf:"policy,omitempty"`
 
 	// Number of days that AWS Secrets Manager waits before it can delete the secret. This value can be 0 to force deletion without recovery or range from 7 to 30 days. The default value is 30.
-	RecoveryWindowInDays *float64 `json:"recoveryWindowInDays,omitempty" tf:"recovery_window_in_days,omitempty"`
+	RecoveryWindowInDays *int64 `json:"recoveryWindowInDays,omitempty" tf:"recovery_window_in_days,omitempty"`
 
 	// Configuration block to support secret replication. See details below.
 	Replica []ReplicaObservation `json:"replica,omitempty" tf:"replica,omitempty"`
@@ -159,7 +159,7 @@ type SecretParameters struct {
 
 	// Number of days that AWS Secrets Manager waits before it can delete the secret. This value can be 0 to force deletion without recovery or range from 7 to 30 days. The default value is 30.
 	// +kubebuilder:validation:Optional
-	RecoveryWindowInDays *float64 `json:"recoveryWindowInDays,omitempty" tf:"recovery_window_in_days,omitempty"`
+	RecoveryWindowInDays *int64 `json:"recoveryWindowInDays,omitempty" tf:"recovery_window_in_days,omitempty"`
 
 	// Region for replicating the secret.
 	// Region is the region you'd like your resource to be created in.
